Replace tocify with tocbot
Tocbot is another library that creates a TOC at runtime, but in
contrast to tocify, it uses native DOM methods and does not
depend on jQuery UI.

Adjust CSS and render the links in the TOC in gray to tone down
this section.

Do not use href attributes in the navigation links on the left,
as this lead to jumping to anchors after changing the library.

[1] https://tscanlin.github.io/tocbot/