Automate update process
To make the update process quicker and less error-prone add an script
"update.js" that does all the previous manual steps automatically. This
includes running jekyll and checking for broken links.

check-links.js is updated so it can be executed standalone or as part of
update.js

Gemfile is updated to not print a warning on windows.